+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 + New `switch` widget type, for composing composite widgets by dynamically
   switching over other types and configurations of widgets. (#820)
++ New `<basic-widget>` directive, refactored out from the inline template in
+  `<widget-card>`. Not a breaking change: `<widget-card>` behaves the same as
+  before, its implementation now delegates to `<basic-widget>` in the basic
+  widget case. (#820)
 
 ### Changed